Refactor manage.web into a package
The single 1684-line manage/web.py grew unwieldy.  Split into
six focused submodules under manage/web/:

- app.py     — FastAPI instance only (separated so route modules
               can import it without a circular dep on __init__).
- state.py   — module-level config (MANAGER, DOMAINS, caches),
               status state machine, lifespan.
- diff.py    — pure per-file diff rendering with the inline cap.
               Used by both the thread page and the review page.
- review.py  — /thread/{tid}/review page + GET/POST routes +
               _format_review_message.
- threads.py — index + thread page + the routes that drive them
               (message, capture, merge, delete).  Owns
               _process_message and _initialize_thread.
- evals.py   — /evals and /evals/run/{id} pages + routes.

manage/web/__init__.py imports the route submodules (which
register their endpoints on `app`) and re-exports the public
names that external callers historically reached via
`from manage.web import X`.  `uvicorn manage.web:app` still
resolves to the same wired-up application.

Test changes: TestPostReviewRoute previously patched
`web._process_message` and `web._get_domain_manager`; route
handlers now look those names up via the modules where they
live (threads / state), so the patches target
`manage.web.threads._process_message` and
`manage.web.review._get_domain_manager` directly.

Also addresses PR #92 review comment about Pygments — confirmed
no first-party callers remain after the diff rewrite; the two
requirements.txt entries are lockfile-pinned transitive deps
(via rich, pytest, ipython, ipython_pygments_lexers) and stay
as-is.

Tests: 340 passing.
